+++ astropy/wcs/setup_package.py
@@ -20,7 +20,7 @@ from astropy_helpers.distutils_helpers import get_distutils_build_option
 from astropy.extern import six
 
 WCSROOT = os.path.relpath(os.path.dirname(__file__))
-WCSVERSION = "5.10"
+WCSVERSION = "5.11"
 
 
 def b(s):