+++ cextern/wcslib/C/flexed/wcsutrn.c
@@ -3984,7 +3984,7 @@ char *wcsutrntext;
 #line 1 "wcsutrn.l"
 /*============================================================================
 
-  WCSLIB 5.10 - an implementation of the FITS WCS standard.
+  WCSLIB 5.11 - an implementation of the FITS WCS standard.
   Copyright (C) 1995-2015, Mark Calabretta
 
   This file is part of WCSLIB.
@@ -4006,7 +4006,7 @@ char *wcsutrntext;
 
   Author: Mark Calabretta, Australia Telescope National Facility, CSIRO.
   http://www.atnf.csiro.au/people/Mark.Calabretta
-  $Id: wcsutrn.c,v 5.10 2015/10/09 08:19:15 mcalabre Exp $
+  $Id: wcsutrn.c,v 5.11 2015/10/18 09:13:06 mcalabre Exp $
 *=============================================================================
 *
 * wcsutrn.l is a Flex description file containing the definition of a lexical